requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
tqdm
scikit-learn
